add a disjoint constraint
The constraint makes sure that intervals from a given set do not
intersect with each other. This can be used to very concisely model some
scheduling problems. For example the flow shop problem. It can also
solve some of the larger flow show instances, e.g., the tai500_20_1.lp
instance.

The propagation algorithm is a straightforward extension of Hall
interval based algorithms for distinct constraints. In it's current
form, the runtime of the algorithm is O(n*n) but can probably be reduced
to O(n*log(n)). Maybe it is also possible to detect some cases where the
algorithm does not have to be run at all (one such case is already noted
in the source code).